(#142) Update Latest Detail Widget JS
* moved two handlebars helpers in LatestMediaWidget.js into handlebars folder
* Renamed handlebars export function to latestMediaWidget_handlebars
* Updated digitalworkplace.custom.js import to reflect new handlebars name
* Removed window.InitializeSlider
* moved remaining LatestMediaWidget.js code into callbacks folder
* Renamed callbacks export function to latestMediaWidget_callbacks
* Added new import for latest media callbacks into digitalworkplace.custom.js
* Updated config

Closes #142
